Turn texture format enum class to enum

+++ Engine/core/Texture.cpp
@@ -89,7 +89,7 @@ void Texture::InitializeTextureDiffuse(const char* texPath)
         return;
     }
 
-    glTexImage2D(GL_TEXTURE_2D, 0, static_cast<GLint>(format), width, height, 0, static_cast<GLint>(format), GL_UNSIGNED_BYTE, data);
+    glTexImage2D(GL_TEXTURE_2D, 0, static_cast<GLint>(format), width, height, 0, format, GL_UNSIGNED_BYTE, data);
     glGenerateMipmap(GL_TEXTURE_2D);
 
     stbi_image_free(data);
